avanzes

+++ oaCore/source/oaStaticMesh.cpp
@@ -8,7 +8,7 @@ StaticMesh::create()
 {
   Mesh::create();
 
-  m_vertexB->init(m_vertices.data(),sizeof(Vertex),static_cast<uint32>(m_vertices.size()));
+  m_vertexB->init(m_vertices.data(),sizeof(Vertex),m_vertices.size());
 }
 
 }
